+++ .idea/compiler.xml
@@ -7,14 +7,15 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="fight" />
         <module name="gamebody" />
         <module name="server" />
         <module name="gateway" />
+        <module name="fighting" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
       <module name="fight" target="11" />
+      <module name="fighting" target="11" />
       <module name="gamebody" target="11" />
       <module name="gateway" target="11" />
       <module name="server" target="11" />
@@ -23,6 +24,7 @@
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
       <module name="fight" options="-parameters" />
+      <module name="fighting" options="-parameters" />
       <module name="gamebody" options="-parameters" />
       <module name="gateway" options="-parameters" />
       <module name="server" options="-parameters" />
