+++ .idea/encodings.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$/fight/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/fighting/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/gamebody/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/gateway/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/server/src/main/java" charset="UTF-8" />
