+++ .idea/misc.xml
@@ -10,6 +10,7 @@
         <option value="$PROJECT_DIR$/gamebody/pom.xml" />
         <option value="$PROJECT_DIR$/gateway/pom.xml" />
         <option value="$PROJECT_DIR$/actions/pom.xml" />
+        <option value="$PROJECT_DIR$/fighting/pom.xml" />
       </list>
     </option>
   </component>
