--- fight/src/main/java/com/getulus/charondor/actions/ExceptionLog.java
@@ -1,16 +1,0 @@
-package com.getulus.charondor.actions;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-
-
-@Configuration
-@Data
-@NoArgsConstructor
-public class ExceptionLog {
-
-    public void log(Exception e) {
-        System.out.println(e);
-    }
-}
