--- fight/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.application.name=actions
-server.port=8083
-eureka.client.service-url.default-zone=http://localhost:8761/eureka
