--- fight/src/test/java/com/getulus/charondor/actions/ActionsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.getulus.charondor.actions;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ActionsApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
