+++ fighting/src/main/java/com/getulus/charondor/fighting/FightingApplication.java
@@ -1,0 +1,13 @@
+package com.getulus.charondor.fighting;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class FightingApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(FightingApplication.class, args);
+    }
+
+}
