+++ fighting/src/test/java/com/getulus/charondor/fighting/FightingApplicationTests.java
@@ -1,0 +1,13 @@
+package com.getulus.charondor.fighting;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class FightingApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
